Move `PatchTestNet` to `NBitcoinHelpers` to avoid CA2255 warning
https://learn.microsoft.com/en-us/dotnet/fundamentals/code-analysis/quality-rules/ca2255
