+++ WalletWasabi.Daemon/ModuleInitializer.cs
@@ -1,0 +1,13 @@
+namespace WalletWasabi.Daemon;
+
+using System.Runtime.CompilerServices;
+using WalletWasabi.Helpers;
+
+class ModuleInitializer
+{
+	[ModuleInitializer]
+	internal static void PatchTestNet()
+	{
+		NBitcoinHelpers.PatchTestNet();
+	}
+}
